+++ CHANGELOG.md
@@ -13,5 +13,6 @@ the [releases page](https://github.com/Consensys/teku/releases).
 - By default, Teku won't allow syncing from genesis, users should use `--checkpoint-sync-url` when starting a new node. It is possible to revert back to the previous behaviour using the flag `--ignore-weak-subjectivity-period-enabled`.
 
 ### Additions and Improvements
+- Support to new Beacon APIs `publishBlindedBlockV2` and `publishBlockV2` which introduce broadcast validation parameter. 
 
 ### Bug Fixes